+++ programs/local/LocalServer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Client/ClientBase.h>
+#include <Client/ClientApplicationBase.h>
 #include <Client/LocalConnection.h>
 
 #include <Core/ServerSettings.h>
@@ -21,7 +21,7 @@ namespace DB
 /// Lightweight Application for clickhouse-local
 /// No networking, no extra configs and working directories, no pid and status files, no dictionaries, no logging.
 /// Quiet mode by default
-class LocalServer : public ClientBase, public Loggers
+class LocalServer : public ClientApplicationBase, public Loggers
 {
 public:
     LocalServer() = default;
